requirements.txt
tensorflow-gpu==1.13.1
pillow==6.0.0
scikit-image==0.15.0
scipy==1.1.0
matplotlib